refactor(channel): split http.rs into core/http.rs and runtime/http.rs
Split channel/http.rs into two files following the core/runtime separation:
- core/http.rs: pure HTTP functionality (ChannelOutboundHttpPolicy, build_outbound_http_client,
  validate_outbound_http_target, read_json_or_text_response, response_body_detail,
  redact_endpoint_status_url, redact_generic_webhook_status_url) and tests
- runtime/http.rs: config binding only (outbound_http_policy_from_config)

channel/http.rs now re-exports from both locations for backward compatibility.
Added mod core; and mod runtime; to channel/mod.rs to enable the module structure.
