+++ crates/app/src/channel/mod.rs
@@ -12,7 +12,9 @@ mod email;
 pub mod feishu;
 #[cfg(feature = "channel-google-chat")]
 mod google_chat;
+mod core;
 mod http;
+mod runtime;
 #[cfg(feature = "channel-imessage")]
 mod imessage;
 #[cfg(feature = "channel-irc")]
