+++ crates/app/src/channel/runtime/mod.rs
@@ -1,16 +1,3 @@
 // crates/app/src/channel/runtime/mod.rs
 
 pub(in crate::channel) mod http;
-pub(in crate::channel) mod state;
-pub(in crate::channel) mod serve;
-pub(in crate::channel) mod turn_feedback;
-pub(in crate::channel) mod types;
-
-pub use serve::{
-    with_channel_serve_runtime, with_channel_serve_runtime_in_dir,
-    with_channel_serve_runtime_with_stop, with_channel_serve_runtime_with_stop_in_dir,
-    ChannelServeCommandSpec, ChannelServeRuntimeSpec, ChannelServeStopHandle,
-};
-pub use state::{ChannelOperationRuntime, ChannelOperationRuntimeTracker};
-pub use turn_feedback::{ChannelTurnFeedbackCapture, ChannelTurnFeedbackPolicy};
-pub use types::{process_channel_batch, ChannelAdapter, ChannelAdapterFeedback};
